api/requirements.txt
tensorflow==2.20.0
tensorflow-serving-api==2.19.1
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
requests
python-dotenv
